Rename `inv`/`inv_mod` modules to `invert`/`invert_mod`
