+++ src/int.rs
@@ -22,7 +22,7 @@ mod div_uint;
 mod encoding;
 mod from;
 mod gcd;
-mod inv_mod;
+mod invert_mod;
 mod mul;
 mod mul_uint;
 mod neg;
