+++ src/uint/invert_mod.rs
